Address code review feedback
1. Remove potential top bar elevation in success screen
2. Hide Stripe logo by default on first pane
3. Move Flow extensions to their own file
